03-asynchronous-request-reply-pattern/brian/webhook/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
line-bot-sdk>1.1.0,<2.0.0
httpx>=0.24.0,<0.25.0
